clean up rakefile

+++ Rakefile
@@ -10,7 +10,7 @@ def configuration
   "Debug"
 end
 
-def test_targets
+def test_platforms
   [
     :ios,
     :tvos
@@ -18,34 +18,29 @@ def test_targets
 end
 
 def schemes
-  {
-    ios: 'Charts',
-    tvos: 'Charts',
-    osx: 'Charts'
-  }
-end
-
-def sdks
-  {
-    ios: 'iphonesimulator',
-    osx: 'macosx',
-    tvos: 'appletvsimulator'
-  }
+  [
+    "Charts"
+    # "ChartsRealm"
+  ]
 end
 
 def devices
   {
-    ios: "name='iPhone 7'",
-    osx: "arch='x86_64'",
-    tvos: "name='Apple TV 1080p'"
-  }
-end
-
-# UUIDs of the Xcode 8 simulators from Travis CI
-def uuids
-  {
-    ios: '5F911B30-5F23-403B-9697-1DFDC24773C8', # iPhone 7
-    tvos: '273D776F-196E-4F2A-AEF2-E1E3EAE99B47' # Apple TV 1080p
+    ios: { 
+      sdk: "iphonesimulator", 
+      device: "name='iPhone 7'", 
+      uuid: "5F911B30-5F23-403B-9697-1DFDC24773C8" 
+    },
+    macos: { 
+      sdk: "macosx", 
+      device: "arch='x86_64'", 
+      uuid: nil 
+    },
+    tvos: { 
+      sdk: "appletvsimulator", 
+      device: "name='Apple TV 1080p'", 
+      uuid: "273D776F-196E-4F2A-AEF2-E1E3EAE99B47" 
+    }
   }
 end
 
@@ -70,32 +65,46 @@ def xcodebuild(type, name, scheme, configuration, sdk, destination, tasks, xcpre
 
 end
 
-def execute(tasks, platform, xcprety_args)
+def run_xcodebuild(schemes, tasks, destination, is_test, xcprety_args)
+  sdk = destination[:sdk]
+  device = destination[:device]
+  uuid = destination[:uuid]
 
-  # platform specific settings
-  sdk = sdks[platform]
-  scheme = schemes[platform]
-  destination = devices[platform]
-
-  open_simulator_and_sleep(uuids[platform])
-
-  # check if xcodebuild needs to be run on multiple devices
-  if destination.respond_to?('map')
-    destination.map do |destination|
-      xcodebuild type, project_name, scheme, configuration, sdk, destination, tasks, xcprety_args
-    end
-  else
-    xcodebuild type, project_name, scheme, configuration, sdk, destination, tasks, xcprety_args
+  if is_test
+    open_simulator_and_sleep uuid
   end
 
+  schemes.each do |scheme| 
+    xcodebuild type, project_name, scheme, configuration, sdk, device, tasks, xcprety_args
+  end
+
+  if is_test
+    sh "killall Simulator"
+  end
 end
 
-desc 'Build, then run tests.'
+def execute(tasks, platform, xcprety_args)
+
+  is_test = tasks.include?("test")
+
+  # platform specific settings
+  destination = devices[platform]
+
+  # check if xcodebuild needs to be run on multiple devices
+  if destination.is_a?(Array)
+    destination.each do |destination|
+        run_xcodebuild schemes, tasks, destination, is_test, xcprety_args
+    end
+  else 
+    run_xcodebuild schemes, tasks, destination, is_test, xcprety_args
+  end
+end
+
+desc "Build, then run tests."
 task :test do
 
-  test_targets.map do |platform|
-    execute 'build test', platform, xcprety_args: '--test'
-    sh "killall Simulator"
+  test_platforms.each do |platform|
+    execute "build test", platform, xcprety_args: "--test"
   end
 
 end
